+++ src/main/java/co/fineants/price/domain/stockprice/client/StockPriceWebSocketHandler.java
@@ -1,4 +1,4 @@
-package co.fineants.price.domain.stock_price.client;
+package co.fineants.price.domain.stockprice.client;
 
 import org.jetbrains.annotations.NotNull;
 import org.springframework.stereotype.Component;
