+++ src/main/java/co/fineants/price/domain/stockprice/service/StockPriceService.java
@@ -1,10 +1,10 @@
-package co.fineants.price.domain.stock_price.service;
+package co.fineants.price.domain.stockprice.service;
 
 import java.util.Set;
 
 import org.springframework.stereotype.Service;
 
-import co.fineants.price.domain.stock_price.repository.StockPriceRepository;
+import co.fineants.price.domain.stockprice.repository.StockPriceRepository;
 import lombok.RequiredArgsConstructor;
 
 @Service
